requirements/base.txt
django-simple-captcha
django_widget_tweaks
django_staticfiles
